Robert - Reorganized the code and got it all working again.
