+++ client/Components/Chart/ForceGraphGenerator.js
@@ -1,4 +1,3 @@
-import React from "react"
 import ReactDOM from "react-dom"
 import * as d3 from "d3"
